requirements.txt
# ==== Core framework ====
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

# ==== Database / Storage ====
sqlalchemy==2.0.35
alembic==1.13.2
sqlite-utils==3.36
psycopg2-binary==2.9.9          # optional for Postgres; safe to leave installed

# ==== File handling & utilities ====
aiofiles==24.1.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
pathlib==1.0.1
tqdm==4.66.5

# ==== LLM / Gemini integration ====
google-generativeai==0.8.3       # Gemini API client
langchain==0.3.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1

# ==== Reinforcement Learning Core ====
torch==2.4.1
numpy==1.26.4
gymnasium==0.29.1
matplotlib==3.9.2
pandas==2.2.3

# ==== Self-Healing / Test Execution ====
pytest==8.3.3
# use built-in subprocess; removed subprocess-run
# removed postman-api — use Newman CLI instead

# ==== Logging & Visualization ====
rich==13.9.2
mlflow==2.16.0
tensorboard==2.17.0

# ==== Optional helpers ====
fastapi-utils==0.7.0
aiohttp==3.10.5
